refactor: Move base HRRR region job tests to new location
